+++ src/integrationTest/java/io/bitgrillr/gocddockerexecplugin/IntegrationTest.java
@@ -1,10 +1,10 @@
-package io.bitgrillr.godockerbuildplugin;
+package io.bitgrillr.gocddockerexecplugin;
 
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 
-import io.bitgrillr.godockerbuildplugin.utils.GoTestUtils;
+import io.bitgrillr.gocddockerexecplugin.utils.GoTestUtils;
 import org.junit.Test;
 
 public class IntegrationTest {