+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ## 1.0.0
 
-* [3](https://github.com/cma-arnold/go-docker-build-plugin/issues/3) - Get the plugin working with docker.
-* [2](https://github.com/cma-arnold/go-docker-build-plugin/issues/2) - Create inital plugin build.
+* [3](https://github.com/cma-arnold/gocd-docker-exec-plugin/issues/3) - Get the plugin working with docker.
+* [2](https://github.com/cma-arnold/gocd-docker-exec-plugin/issues/2) - Create inital plugin build.